+++ great_expectations/datasource/data_connector/file_path_data_connector.py
@@ -73,6 +73,7 @@ class FilePathDataConnector(DataConnector):
 
         if default_regex is None:
             default_regex = {}
+
         self._default_regex = default_regex
 
         self._sorters = build_sorters_from_config(config_list=sorters)  # type: ignore[arg-type]
@@ -181,7 +182,6 @@ class FilePathDataConnector(DataConnector):
             )
 
         if batch_request.data_connector_query is not None:
-
             data_connector_query_dict = batch_request.data_connector_query.copy()
             if (
                 batch_request.limit is not None
@@ -216,6 +216,7 @@ class FilePathDataConnector(DataConnector):
             batch_definition_list = sorter.get_sorted_batch_definitions(
                 batch_definitions=batch_definition_list
             )
+
         return batch_definition_list
 
     def _map_data_reference_to_batch_definition_list(
@@ -261,19 +262,6 @@ class FilePathDataConnector(DataConnector):
         )
         return PathBatchSpec(batch_spec)
 
-    @staticmethod
-    def sanitize_prefix(text: str) -> str:
-        """
-        Takes in a given user-prefix and cleans it to work with file-system traversal methods
-        (i.e. add '/' to the end of a string meant to represent a directory)
-        """
-        _, ext = os.path.splitext(text)
-        if ext:
-            # Provided prefix is a filename so no adjustment is necessary
-            return text
-        # Provided prefix is a directory (so we want to ensure we append it with '/')
-        return os.path.join(text, "")
-
     def _generate_batch_spec_parameters_from_batch_definition(
         self, batch_definition: BatchDefinition
     ) -> dict:
@@ -286,9 +274,11 @@ class FilePathDataConnector(DataConnector):
 batch identifiers {batch_definition.batch_identifiers} from batch definition {batch_definition}.
 """
             )
+
         path = self._get_full_file_path(
             path=path, data_asset_name=batch_definition.data_asset_name
         )
+
         return {"path": path}
 
     def _validate_batch_request(self, batch_request: BatchRequestBase) -> None:
@@ -312,6 +302,7 @@ batch identifiers {batch_definition.batch_identifiers} from batch definition {ba
 configured group_name.
                     """
                 )
+
             if len(group_names) < len(self.sorters):
                 raise gx_exceptions.DataConnectorError(
                     f"""DataConnector "{self.name}" is configured with {len(group_names)} group names;
